renames the SelectInput base component to Select
