+++ src/components/Select/index.js
@@ -1,0 +1,3 @@
+import Select from './Select';
+
+export default Select;
