+++ src/signals/incident/components/form/DateTimeInput/index.js
@@ -3,7 +3,7 @@ import format from 'date-fns/format';
 import subDays from 'date-fns/subDays';
 import PropTypes from 'prop-types';
 import nl from 'date-fns/locale/nl';
-import Select from 'components/SelectInput';
+import Select from 'components/Select';
 import { capitalize } from 'shared/services/date-utils';
 import Header from '../Header';
